+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * feature: Redesign flow_collectors.php for formalize support for IPFIX in Flowview
 * feature: Store peers that connect to listeners for review and tagging
 * feature: Inform the user of how many listener streams are present and their status
+* feature: Track detected template definitions in the database and provide a way to view them
 * feature#97: Add FreeBSD service control
 
 --- 3.3 ---